// src/modules/slug/types.ts
export namespace SlugData {
  export interface Slug {
    schema: any;
    articleSection: string;
    date: string;
    title: string;
    url: any;
    id?: number;
    minutes: string;
    link: string;
    slug: string;
    content: string;
    twitter_img: string;
    description: string;
    author: string;
    avatarUrl: string;
  }
}